refactor: move applyPrecipitation to its own file
