+++ src/game-logic/reducers/index.test.js
@@ -1,10 +1,5 @@
 import { shapeOf, testCrop, testItem } from '../../test-utils'
-import {
-  OUT_OF_COW_FEED_NOTIFICATION,
-  RAIN_MESSAGE,
-  STORM_MESSAGE,
-  STORM_DESTROYS_SCARECROWS_MESSAGE,
-} from '../../strings'
+import { OUT_OF_COW_FEED_NOTIFICATION } from '../../strings'
 import {
   ACHIEVEMENT_COMPLETED,
   COW_ATTRITION_MESSAGE,
@@ -462,114 +457,6 @@ describe('computeStateForNextDay', () => {
     expect(firstRow[0].daysWatered).toBe(1)
     expect(firstRow[0].daysOld).toBe(1)
     expect(todaysNotifications).toBeEmpty()
-  })
-})
-
-describe('applyPrecipitation', () => {
-  test('waters all plots', () => {
-    const state = fn.applyPrecipitation({
-      field: [
-        [
-          testCrop({
-            wasWateredToday: false,
-          }),
-          testCrop({
-            wasWateredToday: false,
-          }),
-        ],
-      ],
-      inventory: [],
-      newDayNotifications: [],
-    })
-
-    expect(state.field[0][0].wasWateredToday).toBe(true)
-    expect(state.field[0][1].wasWateredToday).toBe(true)
-  })
-
-  describe('rain shower', () => {
-    test('waters all plots', () => {
-      jest.spyOn(Math, 'random').mockReturnValue(1)
-      const state = fn.applyPrecipitation({
-        field: [[]],
-        inventory: [],
-        newDayNotifications: [],
-      })
-
-      expect(state.newDayNotifications[0]).toEqual({
-        message: RAIN_MESSAGE,
-        severity: 'info',
-      })
-    })
-  })
-
-  describe('storm', () => {
-    beforeEach(() => {
-      jest.spyOn(Math, 'random').mockReturnValue(0)
-    })
-
-    describe('scarecrows are planted', () => {
-      test('scarecrows are destroyed', () => {
-        const state = fn.applyPrecipitation({
-          field: [[getPlotContentFromItemId(SCARECROW_ITEM_ID)]],
-          inventory: [],
-          newDayNotifications: [],
-        })
-
-        expect(state.field[0][0]).toBe(null)
-        expect(state.newDayNotifications[0]).toEqual({
-          message: STORM_DESTROYS_SCARECROWS_MESSAGE,
-          severity: 'error',
-        })
-      })
-
-      describe('scarecows are rainbow fertilized', () => {
-        test('scarecrows are replaced based on available inventory', () => {
-          const { field, inventory } = fn.applyPrecipitation({
-            field: [
-              [
-                {
-                  ...getPlotContentFromItemId(SCARECROW_ITEM_ID),
-                  fertilizerType: fertilizerType.RAINBOW,
-                },
-                {
-                  ...getPlotContentFromItemId(SCARECROW_ITEM_ID),
-                  fertilizerType: fertilizerType.RAINBOW,
-                },
-              ],
-            ],
-            inventory: [{ id: 'scarecrow', quantity: 1 }],
-            newDayNotifications: [],
-          })
-
-          // Scarecrow is replanted from inventory
-          expect(field[0][0]).toEqual({
-            ...getPlotContentFromItemId(SCARECROW_ITEM_ID),
-            fertilizerType: fertilizerType.RAINBOW,
-          })
-
-          // Scarecrow replacement was not available
-          expect(field[0][1]).toBe(null)
-
-          // Scarecrow inventory is consumed
-          expect(inventory).toEqual([])
-        })
-      })
-    })
-
-    describe('scarecrows are not planted', () => {
-      test('shows appropriate message', () => {
-        const state = fn.applyPrecipitation({
-          field: [[]],
-          inventory: [],
-          newDayNotifications: [],
-        })
-
-        expect(state.newDayNotifications[0]).toEqual({
-          message: STORM_MESSAGE,
-          severity: 'info',
-        })
-      })
-    })
   })
 })
